fix: seed separation
