+++ core/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod entity;
 pub mod entity_trait;
+pub mod seed;
 pub mod slug_config;
 pub mod utils;
